Improved Index.html + new CSS file
The 'About' section has been removed since the rest of the page explains what the mod is about in much more detail. The new stylesheet is designed to improve the table shown near the bottom of the page which lists all of the changes to MoreGDT over time, the new table is much cleaner and fits the screen better than the previous one did along with having better sized columns so the text doesn't take up two lines when it isn't necessary.